+++ jobbergate-cli/CHANGELOG.md
@@ -4,6 +4,7 @@ This file keeps track of all notable changes to jobbergate-cli
 
 ## Unreleased
 
+- Modified internal details to ensure commands that require no authentication face no configuration errors [ASP-5745]
 
 ## 5.3.0 -- 2024-09-09