+++ jobbergate-cli/CHANGELOG.md
@@ -5,6 +5,7 @@ This file keeps track of all notable changes to jobbergate-cli
 ## Unreleased
 
 
+## 5.4.0a1 -- 2024-11-05
 ## 5.4.0a0 -- 2024-11-04
 - Enabled positional arguments to select entries using the cli [ASP-5649]
 - Modified internal details to ensure commands that require no authentication face no configuration errors [ASP-5745]